Refactor array package as a standard go package practice
