+++ plugin/nube/protocals/lorawan/utils.go
@@ -1,19 +1,19 @@
 package main
 
 import (
-	"github.com/NubeIO/flow-framework/utils"
+	"github.com/NubeIO/flow-framework/utils/array"
 	"strings"
 )
 
 func getPointAddr(s string) (objType, addr string) {
-	mArr := utils.NewArray()
+	arr := array.NewArray()
 	ss := strings.Split(s, "-")
 	for _, e := range ss {
 		if e != "" {
-			mArr.Add(e)
+			arr.Add(e)
 		}
 	}
-	o := mArr.Get(0)
-	a := mArr.Get(1)
+	o := arr.Get(0)
+	a := arr.Get(1)
 	return o.(string), a.(string)
 }
